Issue #3 - add Verilog support is work in progress
